refactor(runner): move runners into dedicated directories
