+++ sln/src/DotNetTestNSpec/DesignTime/DesignTimeRunner.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace DotNetTestNSpec
+namespace DotNetTestNSpec.DesignTime
 {
     public class DesignTimeRunner : ITestRunner
     {
